data/regular: fix monitor ids (again)

+++ src/data/regular.rs
@@ -94,7 +94,7 @@ pub enum Transforms {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Monitor {
     /// The monitor id
-    pub id: i64,
+    pub id: i128,
     /// The monitor's name
     pub name: String,
     /// The monitor's description
@@ -217,7 +217,7 @@ pub struct Client {
     #[serde(rename = "fullscreenMode")]
     pub fullscreen_mode: i8,
     /// The monitor id the window is on
-    pub monitor: i64,
+    pub monitor: i128,
     /// The initial window class
     #[serde(rename = "initialClass")]
     pub initial_class: String,
